pyproject.toml
[tool.poetry]
name = "tbot"
version = "0.1.0"
description = ""
authors = ["Matvey <matveykawork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
colorama = "^0.4.6"
redis = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
